+++ CHANGELOG.md
@@ -14,6 +14,8 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 * Removed
 * Fixed
   * bug fix, hen loaded an old project will less columns settings than expected
+  * ODS reader: detects if cell repeats more that 25 times
+  * XLS reader: converts float cell to a string
 
 ## 1.2.0 - 2025-01-17